Add files via upload
This Java program generates a random number between 1 and 100 and prompts the user to guess it within 10 attempts. If the user guesses the number correctly, the program congratulates them and displays the number of attempts it took to guess the number. The program also keeps track of the user’s score by adding the number of attempts to guess the number to the user’s score. If the user fails to guess the number within 10 attempts, the program displays the correct number and prompts the user to play again. The program ends when the user chooses not to play again