Add files via upload
This C program calculates the total marks, average percentage, and grade based on the average percentage achieved. It prompts the user to input the marks obtained in each subject (out of 100) and then calculates the total marks by summing up the marks obtained in all subjects. The average percentage is calculated by dividing the total marks by the total number of subjects. The program then assigns grades based on the average percentage achieved using a predefined table. Finally, the program displays the total marks, average percentage, and the corresponding grade to the user.